yiyu
diary web

+++ controllers/tag_controller.go
@@ -1,6 +1,7 @@
 package controllers
 
 import (
+	"awesomeProject1/middleware"
 	"awesomeProject1/pkg/models"
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -20,7 +21,8 @@ func NewTagController(db *gorm.DB) *TagController {
 
 // Routes 注册标签相关的路由
 func (ctrl *TagController) Routes(r *gin.Engine) {
-	tags := r.Group("/tags")
+	//添加jwt校验
+	tags := r.Group("/tags", middleware.JwtMiddleware())
 	{
 		tags.POST("", ctrl.CreateTag)       // 创建标签
 		tags.GET("", ctrl.GetTags)          // 获取标签列表
@@ -38,6 +40,12 @@ func (ctrl *TagController) CreateTag(c *gin.Context) {
 		//	c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		//	return
 		//}
+
+		//判断表中是否存在
+		if ctrl.DB.Where("name = ?", tag.Name).First(&tag).RowsAffected > 0 {
+			c.JSON(http.StatusConflict, gin.H{"message": "Tag already exists"})
+			return
+		}
 		result := ctrl.DB.Create(&tag)
 		if result.Error != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
@@ -118,6 +126,33 @@ func (ctrl *TagController) DeleteTag(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
 		return
 	}
+
+	//查询标签下的日记id
+	var diaryTags []models.DiaryTags
+	tx := ctrl.DB.Find(&diaryTags, "tag_id = ?", tagID)
+	if tx.Error != nil { //删除标签下的日记失败
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete diary under the tag"})
+		return
+	}
+
+	//根据日记id软删除日记
+	var err2 error
+	for _, diaryTag := range diaryTags {
+		var diary models.Diary
+		result2 := ctrl.DB.Where("id = ? AND deleted_at IS NULL", diaryTag.DiaryID).First(&diary)
+		if result2.Error != nil {
+			err2 = result2.Error
+			continue
+		}
+		if err := ctrl.DB.Delete(&models.Diary{}, diaryTag.DiaryID).Error; err != nil {
+			err2 = err
+		}
+	}
+	if err2 != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete diary under the tag"})
+		return
+	}
+
 	result = ctrl.DB.Delete(&tag)
 	if result.Error != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": result.Error.Error()})
